mover/mover.py
"""
Data Mover Service
Consumes move events from Kafka and executes data migration between MinIO tiers
Implements copy-verify-delete pattern for data consistency
"""

import os
import time
import json
from datetime import datetime
from kafka import KafkaConsumer
from minio import Minio
from minio.error import S3Error
from minio.commonconfig import CopySource
import requests
import io
import logging

logger = logging.getLogger(__name__)

# Configuration
KAFKA_BOOTSTRAP_SERVERS = os.getenv("KAFKA_BOOTSTRAP_SERVERS", "kafka:29092")
KAFKA_TOPIC = os.getenv("KAFKA_TOPIC", "move_events")
KAFKA_GROUP_ID = "mover-group"
MINIO_ENDPOINT = os.getenv("MINIO_ENDPOINT", "minio:9000")
MINIO_ACCESS_KEY = os.getenv("MINIO_ACCESS_KEY", "minioadmin")
MINIO_SECRET_KEY = os.getenv("MINIO_SECRET_KEY", "minioadmin123")
BACKEND_URL = os.getenv("BACKEND_URL", "http://backend:8000")

def connect_kafka():
    """Connect to Kafka consumer with retry logic"""
    max_retries = 10
    retry_delay = 3
    
    for attempt in range(max_retries):
        try:
            consumer = KafkaConsumer(
                KAFKA_TOPIC,
                bootstrap_servers=KAFKA_BOOTSTRAP_SERVERS.split(','),
                group_id=KAFKA_GROUP_ID,
                value_deserializer=lambda m: json.loads(m.decode('utf-8')),
                auto_offset_reset='earliest',
                enable_auto_commit=True
            )
            logger.info("Mover connected to Kafka at %s", KAFKA_BOOTSTRAP_SERVERS)
            logger.info("Consuming from topic: %s", KAFKA_TOPIC)
            return consumer
        except Exception as e:
            logger.warning("Kafka connection attempt %s/%s failed: %s", attempt + 1, max_retries, e)
            if attempt < max_retries - 1:
                time.sleep(retry_delay)
            else:
                raise

def connect_minio():
    """Connect to MinIO"""
    max_retries = 10
    retry_delay = 3
    
    for attempt in range(max_retries):
        try:
            client = Minio(
                MINIO_ENDPOINT,
                access_key=MINIO_ACCESS_KEY,
                secret_key=MINIO_SECRET_KEY,
                secure=False
            )
            
            # Test connection by listing buckets
            buckets = client.list_buckets()
            logger.info("Mover connected to MinIO at %s", MINIO_ENDPOINT)
            logger.debug("Available buckets: %s", [b.name for b in buckets])
            return client
        except Exception as e:
            logger.warning("MinIO connection attempt %s/%s failed: %s", attempt + 1, max_retries, e)
            if attempt < max_retries - 1:
                time.sleep(retry_delay)
            else:
                raise

def update_movement_status(movement_id: int, status: str, error_message: str = None, max_retries: int = 3):
    """Update movement status in backend with retry logic"""
    for attempt in range(max_retries):
        try:
            params = {"status": status}
            if error_message:
                params["error_message"] = error_message
            
            response = requests.put(
                f"{BACKEND_URL}/movements/{movement_id}/status",
                params=params,
                timeout=10
            )
            
            if response.status_code == 200:
                logger.info("Updated movement %s status to %s", movement_id, status)
                return True
            else:
                logger.warning(
                    "Failed to update movement status (HTTP %s)", response.status_code
                )
                if attempt < max_retries - 1:
                    time.sleep(2 ** attempt)  # Exponential backoff
        except Exception as e:
            logger.error(
                "Failed to update movement status (attempt %s/%s): %s",
                attempt + 1,
                max_retries,
                e
            )
            if attempt < max_retries - 1:
                time.sleep(2 ** attempt)  # Exponential backoff: 1s, 2s, 4s
            else:
                logger.error("Giving up after %s attempts updating movement status", max_retries)
    
    return False

def create_movement_record(dataset_id: int, from_tier: str, to_tier: str, size_bytes: int, reason: str):
    """Create a movement record in backend and return its ID"""
    try:
        # For now, we'll use a simple heuristic
        # In a real system, the backend would return the movement ID
        # Since the movement is already created by the backend's /decide endpoint,
        # we'll just return a placeholder
        return None
    except Exception as e:
        logger.warning("Failed to create movement record: %s", e)
        return None

def get_object_path(dataset_name: str, version: int) -> str:
    """Generate object path with versioning"""
    return f"datasets/{dataset_name}/v{version}/data"

def move_data(
    minio_client: Minio,
    dataset_id: int,
    dataset_name: str,
    from_tier: str,
    to_tier: str,
    size_bytes: int,
    version: int = 1,
    movement_id: int = None
):
    """
    Move data between MinIO buckets using copy-verify-delete pattern
    
    Steps:
    1. Check if source exists
    2. Copy to destination
    3. Verify copy
    4. Update backend
    5. Delete source (optional - kept for safety)
    """
    try:
        logger.info(
            "Starting migration for dataset %s (%s -> %s, %.2f MB)",
            dataset_name,
            from_tier,
            to_tier,
            size_bytes / (1024**2)
        )
        
        source_bucket = from_tier
        dest_bucket = to_tier
        object_name = get_object_path(dataset_name, version)
        
        # Step 1: Check if source exists, if not create dummy data
        try:
            minio_client.stat_object(source_bucket, object_name)
            logger.debug("Source object exists: %s/%s", source_bucket, object_name)
            
            # Step 2: Copy to destination using CopySource
            logger.debug("Copying object %s from %s to %s", object_name, source_bucket, dest_bucket)
            minio_client.copy_object(
                dest_bucket,
                object_name,
                CopySource(source_bucket, object_name)
            )
            
        except S3Error as e:
            if e.code == 'NoSuchKey':
                # Source doesn't exist, create dummy data in destination
                logger.info("Source missing; creating new object in %s", dest_bucket)
                
                # Create dummy data
                data_content = f"Dataset: {dataset_name}\nTier: {to_tier}\nSize: {size_bytes}\nTimestamp: {datetime.utcnow().isoformat()}\n"
                data_bytes = data_content.encode('utf-8')
                data_stream = io.BytesIO(data_bytes)
                
                minio_client.put_object(
                    dest_bucket,
                    object_name,
                    data_stream,
                    length=len(data_bytes),
                    content_type='text/plain'
                )
            else:
                raise
        
        # Step 3: Verify destination
        logger.debug("Verifying destination object %s/%s", dest_bucket, object_name)
        dest_stat = minio_client.stat_object(dest_bucket, object_name)
        logger.debug("Destination verified: %s/%s (size: %s)", dest_bucket, object_name, dest_stat.size)
        
        # Step 4: Update backend status
        if movement_id:
            update_movement_status(movement_id, "completed")
        
        logger.info("Migration completed successfully for dataset %s", dataset_name)
        return True
        
    except Exception as e:
        error_msg = f"Migration failed: {str(e)}"
        logger.error(error_msg)
        
        if movement_id:
            update_movement_status(movement_id, "failed", error_msg)
        
        return False

def process_move_event(minio_client: Minio, event: dict):
    """Process a single move event"""
    try:
        movement_id = event.get('movement_id')  # CRITICAL: Get movement ID
        dataset_id = event.get('dataset_id')
        dataset_name = event.get('dataset_name')
        from_tier = event.get('from_tier')
        to_tier = event.get('to_tier')
        size_bytes = event.get('size_bytes')
        reason = event.get('reason', 'unknown')
        
        logger.info(
            "Processing move event | movement_id=%s dataset=%s (%s) %s -> %s reason=%s",
            movement_id,
            dataset_name,
            dataset_id,
            from_tier,
            to_tier,
            reason
        )
        
        if not movement_id:
            logger.error("Move event missing movement_id; cannot track status")
            return False
        
        # Update status to in_progress
        update_movement_status(movement_id, "in_progress")
        
        # Execute the move
        success = move_data(
            minio_client=minio_client,
            dataset_id=dataset_id,
            dataset_name=dataset_name,
            from_tier=from_tier,
            to_tier=to_tier,
            size_bytes=size_bytes,
            version=1,
            movement_id=movement_id
        )
        
        if success:
            logger.info("Move event processed successfully for dataset %s", dataset_name)
        else:
            logger.error("Move event processing failed for dataset %s", dataset_name)
        
        return success
        
    except Exception as e:
        logger.error("Error processing event: %s", e)
        return False

def main():
    """Main mover loop"""
    logger.info("Starting Data Mover Service")
    
    # Connect to MinIO
    time.sleep(5)
    minio_client = connect_minio()
    
    # Connect to Kafka
    time.sleep(2)
    consumer = connect_kafka()
    
    logger.info("Listening for move events")
    
    try:
        for message in consumer:
            event = message.value
            process_move_event(minio_client, event)
    
    except KeyboardInterrupt:
        logger.info("Shutting down mover (keyboard interrupt)")
    finally:
        consumer.close()

if __name__ == "__main__":
    main()
