mover/requirements.txt
kafka-python==2.0.2
minio==7.2.0
requests==2.31.0


